Updated create project screen to edit projects too
